+++ editor/editor-plugin-block-controls/src/editor-commands/utils/move-node-utils.ts
@@ -20,6 +20,7 @@ export const getCurrentNodePosFromDragHandleSelection = ({
 
 	if (
 		selection.empty &&
+		expValEqualsNoExposure('platform_editor_block_menu', 'isEnabled', true) &&
 		expValEqualsNoExposure('platform_editor_block_menu_empty_line', 'isEnabled', true)
 	) {
 		currentNodePos = selection.$from.pos;
@@ -70,7 +71,10 @@ export const getPosWhenMoveNodeDown = ({
 
 	const nodeAfter = tr.doc.nodeAt(nodeAfterPos);
 
-	if (expValEqualsNoExposure('platform_editor_block_menu_empty_line', 'isEnabled', true)) {
+	if (
+		expValEqualsNoExposure('platform_editor_block_menu', 'isEnabled', true) &&
+		expValEqualsNoExposure('platform_editor_block_menu_empty_line', 'isEnabled', true)
+	) {
 		const nodeAtCurrentPos = tr.doc.nodeAt($currentNodePos.pos);
 		// if move empty line down to another empty line, move to the position of the next empty line
 		if (