+++ editor/editor-plugin-block-type/src/ui/index.tsx
@@ -6,11 +6,13 @@ import {
 	FORMAT_HEADING_5_MENU_ITEM,
 	FORMAT_HEADING_6_MENU_ITEM,
 	FORMAT_NESTED_MENU_RANK,
+	FORMAT_NESTED_MENU_RANK_REVISED,
 	FORMAT_PARAGRAPH_MENU_ITEM,
 	FORMAT_QUOTE_MENU_ITEM,
 } from '@atlaskit/editor-common/block-menu';
 import type { ExtractInjectionAPI } from '@atlaskit/editor-common/types';
 import type { RegisterBlockMenuComponent } from '@atlaskit/editor-plugin-block-menu';
+import { fg } from '@atlaskit/platform-feature-flags';
 
 import type { BlockTypePlugin } from '../blockTypePluginType';
 
@@ -28,7 +30,9 @@ export const getBlockTypeComponents = (
 			parent: {
 				type: 'block-menu-section' as const,
 				key: 'nested-menu-format-section-primary',
-				rank: FORMAT_NESTED_MENU_RANK[FORMAT_HEADING_1_MENU_ITEM.key],
+				rank: fg('platform_editor_block_menu_format_rank_revised')
+					? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_HEADING_1_MENU_ITEM.key]
+					: FORMAT_NESTED_MENU_RANK[FORMAT_HEADING_1_MENU_ITEM.key],
 			},
 			component: createHeadingBlockMenuItem({ level: 1, api }),
 		},
@@ -38,7 +42,9 @@ export const getBlockTypeComponents = (
 			parent: {
 				type: 'block-menu-section' as const,
 				key: 'nested-menu-format-section-primary',
-				rank: FORMAT_NESTED_MENU_RANK[FORMAT_HEADING_2_MENU_ITEM.key],
+				rank: fg('platform_editor_block_menu_format_rank_revised')
+					? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_HEADING_2_MENU_ITEM.key]
+					: FORMAT_NESTED_MENU_RANK[FORMAT_HEADING_2_MENU_ITEM.key],
 			},
 			component: createHeadingBlockMenuItem({ level: 2, api }),
 		},
@@ -49,7 +55,9 @@ export const getBlockTypeComponents = (
 			parent: {
 				type: 'block-menu-section' as const,
 				key: 'nested-menu-format-section-primary',
-				rank: FORMAT_NESTED_MENU_RANK[FORMAT_HEADING_3_MENU_ITEM.key],
+				rank: fg('platform_editor_block_menu_format_rank_revised')
+					? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_HEADING_3_MENU_ITEM.key]
+					: FORMAT_NESTED_MENU_RANK[FORMAT_HEADING_3_MENU_ITEM.key],
 			},
 			component: createHeadingBlockMenuItem({ level: 3, api }),
 		},
@@ -59,7 +67,9 @@ export const getBlockTypeComponents = (
 			parent: {
 				type: 'block-menu-section' as const,
 				key: 'nested-menu-format-section-primary',
-				rank: FORMAT_NESTED_MENU_RANK[FORMAT_HEADING_4_MENU_ITEM.key],
+				rank: fg('platform_editor_block_menu_format_rank_revised')
+					? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_HEADING_4_MENU_ITEM.key]
+					: FORMAT_NESTED_MENU_RANK[FORMAT_HEADING_4_MENU_ITEM.key],
 			},
 			component: createHeadingBlockMenuItem({ level: 4, api }),
 		},
@@ -69,7 +79,9 @@ export const getBlockTypeComponents = (
 			parent: {
 				type: 'block-menu-section' as const,
 				key: 'nested-menu-format-section-primary',
-				rank: FORMAT_NESTED_MENU_RANK[FORMAT_HEADING_5_MENU_ITEM.key],
+				rank: fg('platform_editor_block_menu_format_rank_revised')
+					? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_HEADING_5_MENU_ITEM.key]
+					: FORMAT_NESTED_MENU_RANK[FORMAT_HEADING_5_MENU_ITEM.key],
 			},
 			component: createHeadingBlockMenuItem({ level: 5, api }),
 		},
@@ -79,7 +91,9 @@ export const getBlockTypeComponents = (
 			parent: {
 				type: 'block-menu-section' as const,
 				key: 'nested-menu-format-section-primary',
-				rank: FORMAT_NESTED_MENU_RANK[FORMAT_HEADING_6_MENU_ITEM.key],
+				rank: fg('platform_editor_block_menu_format_rank_revised')
+					? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_HEADING_6_MENU_ITEM.key]
+					: FORMAT_NESTED_MENU_RANK[FORMAT_HEADING_6_MENU_ITEM.key],
 			},
 			component: createHeadingBlockMenuItem({ level: 6, api }),
 		},
@@ -89,7 +103,9 @@ export const getBlockTypeComponents = (
 			parent: {
 				type: 'block-menu-section' as const,
 				key: 'nested-menu-format-section-primary',
-				rank: FORMAT_NESTED_MENU_RANK[FORMAT_PARAGRAPH_MENU_ITEM.key],
+				rank: fg('platform_editor_block_menu_format_rank_revised')
+					? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_PARAGRAPH_MENU_ITEM.key]
+					: FORMAT_NESTED_MENU_RANK[FORMAT_PARAGRAPH_MENU_ITEM.key],
 			},
 			component: createParagraphBlockMenuItem({ api }),
 		},
@@ -99,7 +115,9 @@ export const getBlockTypeComponents = (
 			parent: {
 				type: 'block-menu-section' as const,
 				key: 'nested-menu-format-section-primary',
-				rank: FORMAT_NESTED_MENU_RANK[FORMAT_QUOTE_MENU_ITEM.key],
+				rank: fg('platform_editor_block_menu_format_rank_revised')
+					? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_QUOTE_MENU_ITEM.key]
+					: FORMAT_NESTED_MENU_RANK[FORMAT_QUOTE_MENU_ITEM.key],
 			},
 			component: createQuoteBlockMenuItem({ api }),
 		},