+++ editor/editor-plugin-code-block-advanced/src/codeBlockAdvancedPlugin.tsx
@@ -2,7 +2,9 @@ import {
 	FORMAT_MENU_ITEM,
 	FORMAT_CODE_BLOCK_MENU_ITEM,
 	FORMAT_NESTED_MENU_RANK,
+	FORMAT_NESTED_MENU_RANK_REVISED,
 } from '@atlaskit/editor-common/block-menu';
+import { fg } from '@atlaskit/platform-feature-flags';
 import { expValEqualsNoExposure } from '@atlaskit/tmp-editor-statsig/exp-val-equals-no-exposure';
 
 import type { CodeBlockAdvancedPlugin } from './codeBlockAdvancedPluginType';
@@ -19,7 +21,9 @@ export const codeBlockAdvancedPlugin: CodeBlockAdvancedPlugin = ({ api, config }
 				parent: {
 					type: 'block-menu-section' as const,
 					key: FORMAT_MENU_ITEM.key,
-					rank: FORMAT_NESTED_MENU_RANK[FORMAT_CODE_BLOCK_MENU_ITEM.key],
+					rank: fg('platform_editor_block_menu_format_rank_revised')
+						? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_CODE_BLOCK_MENU_ITEM.key]
+						: FORMAT_NESTED_MENU_RANK[FORMAT_CODE_BLOCK_MENU_ITEM.key],
 				},
 				component: createCodeBlockMenuItem(api),
 			},