+++ editor/editor-plugin-expand/src/singlePlayerExpand/plugin.tsx
@@ -17,6 +17,7 @@ import {
 	FORMAT_MENU_ITEM,
 	FORMAT_EXPAND_MENU_ITEM,
 	FORMAT_NESTED_MENU_RANK,
+	FORMAT_NESTED_MENU_RANK_REVISED,
 } from '@atlaskit/editor-common/block-menu';
 import { toolbarInsertBlockMessages as messages } from '@atlaskit/editor-common/messages';
 import { IconExpand } from '@atlaskit/editor-common/quick-insert';
@@ -49,7 +50,9 @@ export let expandPlugin: ExpandPlugin = ({ config: options = {}, api }) => {
 				parent: {
 					type: 'block-menu-section' as const,
 					key: FORMAT_MENU_ITEM.key,
-					rank: FORMAT_NESTED_MENU_RANK[FORMAT_EXPAND_MENU_ITEM.key],
+					rank: fg('platform_editor_block_menu_format_rank_revised')
+						? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_EXPAND_MENU_ITEM.key]
+						: FORMAT_NESTED_MENU_RANK[FORMAT_EXPAND_MENU_ITEM.key],
 				},
 				component: createExpandBlockMenuItem(api),
 			},