+++ editor/editor-plugin-layout/src/layoutPlugin.tsx
@@ -21,6 +21,7 @@ import {
 	FORMAT_MENU_ITEM,
 	FORMAT_LAYOUT_MENU_ITEM,
 	FORMAT_NESTED_MENU_RANK,
+	FORMAT_NESTED_MENU_RANK_REVISED,
 } from '@atlaskit/editor-common/block-menu';
 import { type EventDispatcher } from '@atlaskit/editor-common/event-dispatcher';
 import {
@@ -101,7 +102,9 @@ export const layoutPlugin: LayoutPlugin = ({ config: options = {}, api }) => {
 				parent: {
 					type: 'block-menu-section' as const,
 					key: FORMAT_MENU_ITEM.key,
-					rank: FORMAT_NESTED_MENU_RANK[FORMAT_LAYOUT_MENU_ITEM.key],
+					rank: fg('platform_editor_block_menu_format_rank_revised')
+						? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_LAYOUT_MENU_ITEM.key]
+						: FORMAT_NESTED_MENU_RANK[FORMAT_LAYOUT_MENU_ITEM.key],
 				},
 				component: createLayoutBlockMenuItem(api),
 			},