+++ editor/editor-plugin-list/src/ui/index.tsx
@@ -1,10 +1,12 @@
 import {
 	FORMAT_BULLETED_LIST_MENU_ITEM,
 	FORMAT_NESTED_MENU_RANK,
+	FORMAT_NESTED_MENU_RANK_REVISED,
 	FORMAT_NUMBERED_LIST_MENU_ITEM,
 } from '@atlaskit/editor-common/block-menu';
 import type { ExtractInjectionAPI } from '@atlaskit/editor-common/types';
 import type { RegisterBlockMenuComponent } from '@atlaskit/editor-plugin-block-menu';
+import { fg } from '@atlaskit/platform-feature-flags';
 
 import type { ListPlugin } from '../listPluginType';
 
@@ -21,7 +23,9 @@ export const getListComponents = (
 			parent: {
 				type: 'block-menu-section' as const,
 				key: 'nested-menu-format-section-primary',
-				rank: FORMAT_NESTED_MENU_RANK[FORMAT_BULLETED_LIST_MENU_ITEM.key],
+				rank: fg('platform_editor_block_menu_format_rank_revised')
+					? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_BULLETED_LIST_MENU_ITEM.key]
+					: FORMAT_NESTED_MENU_RANK[FORMAT_BULLETED_LIST_MENU_ITEM.key],
 			},
 			component: createBulletedListBlockMenuItem({ api }),
 		},
@@ -31,7 +35,9 @@ export const getListComponents = (
 			parent: {
 				type: 'block-menu-section' as const,
 				key: 'nested-menu-format-section-primary',
-				rank: FORMAT_NESTED_MENU_RANK[FORMAT_NUMBERED_LIST_MENU_ITEM.key],
+				rank: fg('platform_editor_block_menu_format_rank_revised')
+					? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_NUMBERED_LIST_MENU_ITEM.key]
+					: FORMAT_NESTED_MENU_RANK[FORMAT_NUMBERED_LIST_MENU_ITEM.key],
 			},
 			component: createNumberedListBlockMenuItem({ api }),
 		},