+++ editor/editor-plugin-panel/src/panelPlugin.tsx
@@ -13,6 +13,7 @@ import {
 	FORMAT_MENU_ITEM,
 	FORMAT_PANEL_MENU_ITEM,
 	FORMAT_NESTED_MENU_RANK,
+	FORMAT_NESTED_MENU_RANK_REVISED,
 } from '@atlaskit/editor-common/block-menu';
 import { insertSelectedItem } from '@atlaskit/editor-common/insert';
 import { blockTypeMessages } from '@atlaskit/editor-common/messages';
@@ -53,7 +54,9 @@ const panelPlugin: PanelPlugin = ({ config: options = {}, api }) => {
 				parent: {
 					type: 'block-menu-section' as const,
 					key: FORMAT_MENU_ITEM.key,
-					rank: FORMAT_NESTED_MENU_RANK[FORMAT_PANEL_MENU_ITEM.key],
+					rank: fg('platform_editor_block_menu_format_rank_revised')
+						? FORMAT_NESTED_MENU_RANK_REVISED[FORMAT_PANEL_MENU_ITEM.key]
+						: FORMAT_NESTED_MENU_RANK[FORMAT_PANEL_MENU_ITEM.key],
 				},
 				component: createPanelBlockMenuItem(api),
 			},