+++ editor/editor-plugin-extension/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-plugin-extension
 
+## 8.2.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 8.2.2
 
 ### Patch Changes