+++ editor/editor-plugin-find-replace/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-plugin-find-replace
 
+## 5.0.7
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 5.0.6
 
 ### Patch Changes