+++ editor/editor-plugin-media/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-plugin-media
 
+## 7.2.6
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 7.2.5
 
 ### Patch Changes