+++ media/media-test-helpers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-test-helpers
 
+## 39.0.5
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 39.0.4
 
 ### Patch Changes